Adapts collection view cells to dark appearance
